Trying to recreate dot files.
mk-graph.rkt wasn't compiling due to name changes in
kcfa-instantiations.rkt; added "^" to "0cfa" to at
least reference something defined in order to compile.

Cut stuff that was needed for the (commented out) system call to dot
to render the .dot file to pdf.

Changed path to be relative to the recently added
benchmark/introspective.sch instead of an absolute path on ianj's file
system.

I tried to set generate-graph? to #t using syntax-parameterize, but I
must be missing something because no dot files are created.

So I'm still stuck trying to recreate those dot files.